Remove duplicate files from the refactoring
