+++ NewDeclarationInQueue/processfiles/cmodelprocess/formulars/cm_interest_formular.py
@@ -6,12 +6,6 @@ from NewDeclarationInQueue.processfiles.table_builders.man_commercial_builder im
 from NewDeclarationInQueue.processfiles.table_builders.man_professional_builder import ManProfessionalBuilder
 from NewDeclarationInQueue.processfiles.table_builders.member_quality_builder import MemberQualityBuilder
 from NewDeclarationInQueue.processfiles.table_builders.table_content_extractors.ocr_extractor import OcrExtractor
-from NewDeclarationInQueue.processfiles.tableobjects.associate import Associate
-from NewDeclarationInQueue.processfiles.tableobjects.contracts import Contracts
-from NewDeclarationInQueue.processfiles.tableobjects.man_commercial import ManCommercial
-from NewDeclarationInQueue.processfiles.tableobjects.man_professional import ManProfessional
-from NewDeclarationInQueue.processfiles.tableobjects.member_quality import MemberQuality
-from NewDeclarationInQueue.processfiles.tableobjects.table_content_extractors import ocr_extractor
 
 
 class CmInterestFormular(CmFormularBase):
