+++ NewDeclarationInQueue/processfiles/cmodelprocess/formulars/cm_wealth_formular.py
@@ -1,5 +1,3 @@
-
-
 from typing import Tuple
 from NewDeclarationInQueue.processfiles.cmodelprocess.formulars.cm_formular_base import CmFormularBase
 from NewDeclarationInQueue.processfiles.process_messages import ProcessMessages
@@ -13,17 +11,7 @@ from NewDeclarationInQueue.processfiles.table_builders.investment_builder import
 from NewDeclarationInQueue.processfiles.table_builders.mobile_builder import MobileBuilder
 from NewDeclarationInQueue.processfiles.table_builders.parcel_table_builder import ParcelTableBuilder
 from NewDeclarationInQueue.processfiles.table_builders.transport_builder import TransportBuilder
-from NewDeclarationInQueue.processfiles.tableobjects.art import Art
-from NewDeclarationInQueue.processfiles.tableobjects.building import Building
-from NewDeclarationInQueue.processfiles.tableobjects.debt import Debt
-from NewDeclarationInQueue.processfiles.tableobjects.finance import Finance
-from NewDeclarationInQueue.processfiles.tableobjects.gift import Gift
-from NewDeclarationInQueue.processfiles.tableobjects.income import Income
-from NewDeclarationInQueue.processfiles.tableobjects.investment import Investment
-from NewDeclarationInQueue.processfiles.tableobjects.mobile import Mobile
-from NewDeclarationInQueue.processfiles.tableobjects.parcel import Parcel
-from NewDeclarationInQueue.processfiles.tableobjects.table_content_extractors.ocr_extractor import OcrExtractor
-from NewDeclarationInQueue.processfiles.tableobjects.transport import Transport
+from NewDeclarationInQueue.processfiles.table_builders.table_content_extractors.ocr_extractor import OcrExtractor
 
 
 class CmWealthFormular(CmFormularBase):
@@ -32,7 +20,7 @@ class CmWealthFormular(CmFormularBase):
     FIELD_INSTITUTION = "Institutie"
     FIELD_ADDRESS = "Domiciliu"
     FIELD_DOCUMENT_DATE = "Date of filling"
-    
+
     TABLE_PARCELS = "Parcels"
     TABLE_BUILDINGS = "Buildings"
     TABLE_TRANSPORT = "Transport"
@@ -43,33 +31,33 @@ class CmWealthFormular(CmFormularBase):
     TABLE_DEBT = "Debt"
     TABLE_GIFT = "Gift"
     TABLE_INCOME = "Income"
-    
+
     FIELD_OTHER_INCOME = "Other net incomes"
 
-    
     def __init__(self):
         pass
-    
-    def identify_all_data(self, config_formular: dict, raw_tables: list, message: ProcessMessages) -> Tuple[dict, dict, ProcessMessages]:
+
+    def identify_all_data(self, config_formular: dict, raw_tables: list,
+                          message: ProcessMessages) -> Tuple[dict, dict, ProcessMessages]:
         message = self.identify_id_data(message)
         json, raw_json, message = self.identify_tables(config_formular, raw_tables, message)
         return json, raw_json, message
-    
+
     def identify_id_data(self, message: ProcessMessages) -> ProcessMessages:
         fields = self.cmformular["fields"]
-        
+
         self.name = self.get_field_value(fields[self.FIELD_NAME])
         self.job_title = self.get_field_value(fields[self.FIELD_JOB_TITLE])
         self.institution = self.get_field_value(fields[self.FIELD_INSTITUTION])
         self.address = self.get_field_value(fields[self.FIELD_ADDRESS])
         self.doc_date = self.get_field_value(fields[self.FIELD_DOCUMENT_DATE])
-        
+
         return message
-        
-        
-    def identify_tables(self, config_formular: dict, raw_tables: list, message: ProcessMessages) -> Tuple[dict, dict, ProcessMessages]:
+
+    def identify_tables(self, config_formular: dict, raw_tables: list,
+                        message: ProcessMessages) -> Tuple[dict, dict, ProcessMessages]:
         fields = self.cmformular["fields"]
-        
+
         json = {}
         raw_json = {}
         try:
@@ -93,15 +81,15 @@ class CmWealthFormular(CmFormularBase):
                 config_formular, fields, raw_tables, json, raw_json, message)
             message, json, raw_json = self.identify_one_table(self.TABLE_INCOME, 'income', lambda x: IncomeBuilder(OcrExtractor()), \
                 config_formular, fields, raw_tables, json, raw_json, message)
-            
+
             json = self.add_finance_extra_to_json(json, fields)
             raw_json = self.add_finance_extra_to_json(raw_json, fields)
         except Exception as exex:
             print("---------Found excception-------", exex)
             message.add_exception('Error reading the model', exex)
-                  
+
         return json, raw_json, message
-    
+
     def add_finance_extra_to_json(self, root: dict, fields) -> dict:
         root['finance_extra_info'] = self.get_field_value(fields[self.FIELD_OTHER_INCOME])
         root[self.FIELD_NAME] = self.name
@@ -109,9 +97,5 @@ class CmWealthFormular(CmFormularBase):
         root[self.FIELD_INSTITUTION] = self.institution
         root[self.FIELD_ADDRESS] = self.address
         root[self.FIELD_DOCUMENT_DATE] = self.doc_date
-        
+
         return root
-        
-        
-    
-
