+++ NewDeclarationInQueue/processfiles/customprocess/formulars/davere_01.py
@@ -1,5 +1,3 @@
-
-
 from typing import Tuple
 from NewDeclarationInQueue.processfiles.customprocess.formulars.davere import DAvere
 from NewDeclarationInQueue.processfiles.customprocess.search_text_line_parameter import SearchTextLineParameter
@@ -13,7 +11,7 @@ from NewDeclarationInQueue.processfiles.tableobjects.income import Income
 from NewDeclarationInQueue.processfiles.tableobjects.investment import Investment
 from NewDeclarationInQueue.processfiles.tableobjects.mobile import Mobile
 from NewDeclarationInQueue.processfiles.tableobjects.parcel import Parcel
-from NewDeclarationInQueue.processfiles.tableobjects.table_content_extractors.ocr_extractor import OcrExtractor
+from NewDeclarationInQueue.processfiles.table_builders.table_content_extractors.ocr_extractor import OcrExtractor
 from NewDeclarationInQueue.processfiles.tableobjects.transport import Transport
 #from NewDeclarationInQueue.processfiles.customprocess.text_with_special_ch import TextWithSpecialCharacters
 
@@ -24,12 +22,12 @@ class Davere01(DAvere):
     """
         Class for a specific formular for Wealth Declaration
     """
-    
+
     def __init__(self, no_of_pages: int):
         self.no_of_pages = no_of_pages
-    
-    
-    def get_parcels(self, config: TableConfigDetail, data: dict, n_page: int, json: dict, message: ProcessMessages) -> Tuple[dict, ProcessMessages,  int]:
+
+    def get_parcels(self, config: TableConfigDetail, data: dict, n_page: int, json: dict,
+                    message: ProcessMessages) -> Tuple[dict, ProcessMessages, int]:
         """
             Get the info from the table of the specific object
         Args:
@@ -44,14 +42,15 @@ class Davere01(DAvere):
         """
         tables, message, end_page_no = self.find_table_in_document_between_lines(data['ocr_form_response'], \
             n_page, self.no_of_pages, config.upper, config.lower, config.header, message)
-        
+
         message, result = self.extract_table_info_to_json('parcels', tables, lambda x: Parcel(OcrExtractor()), message)
         if message.has_errors() or result is not None:
             json['parcels'] = result
-            
+
         return json, message, (end_page_no if end_page_no > 0 else n_page)
-    
-    def get_buildings(self, config: TableConfigDetail, data: dict, n_page: int, json: dict, message: ProcessMessages) -> Tuple[dict, ProcessMessages, int]:
+
+    def get_buildings(self, config: TableConfigDetail, data: dict, n_page: int, json: dict,
+                      message: ProcessMessages) -> Tuple[dict, ProcessMessages, int]:
         """
             Get the info from the table of the specific object
         Args:
@@ -64,27 +63,28 @@ class Davere01(DAvere):
             Tuple[dict, ProcessMessages,  int]: response JSON for the specific object, processing messages 
                                                     and the page number where the table ends
         """
-        
+
         tables, message, end_page_no = self.find_table_in_document_between_lines(data['ocr_form_response'], \
             n_page, self.no_of_pages, config.upper, config.lower, config.header, message)
         message, result = self.extract_table_info_to_json('buildings', tables, lambda x: Building(), message)
         if message.has_errors() or result is not None:
             json['buildings'] = result
-            
+
         return json, message, (end_page_no if end_page_no > 0 else n_page)
-    
-    def get_transport(self, config: TableConfigDetail, data: dict, n_page: int, json: dict, message: ProcessMessages) -> Tuple[dict, ProcessMessages, int]:
+
+    def get_transport(self, config: TableConfigDetail, data: dict, n_page: int, json: dict,
+                      message: ProcessMessages) -> Tuple[dict, ProcessMessages, int]:
         tables, message, end_page_no = self.find_table_in_document_between_lines(data['ocr_form_response'], \
             n_page, self.no_of_pages, config.upper, config.lower, config.header, message)
-        
+
         message, result = self.extract_table_info_to_json('transport', tables, lambda x: Transport(), message)
         if message.has_errors() or result is not None:
             json['transport'] = result
-            
+
         return json, message, (end_page_no if end_page_no > 0 else n_page)
-    
-    
-    def get_art(self, config: TableConfigDetail, data: dict, n_page: int, json: dict, message: ProcessMessages) -> Tuple[dict, ProcessMessages, int]:
+
+    def get_art(self, config: TableConfigDetail, data: dict, n_page: int, json: dict,
+                message: ProcessMessages) -> Tuple[dict, ProcessMessages, int]:
         """
             Get the info from the table of the specific object
         Args:
@@ -97,17 +97,18 @@ class Davere01(DAvere):
             Tuple[dict, ProcessMessages,  int]: response JSON for the specific object, processing messages 
                                                     and the page number where the table ends
         """
-        
+
         tables, message, end_page_no = self.find_table_in_document_between_lines(data['ocr_form_response'], \
             n_page, self.no_of_pages, config.upper, config.lower, config.header, message)
-        
+
         message, result = self.extract_table_info_to_json('art', tables, lambda x: Art(), message)
         if message.has_errors() or result is not None:
             json['art'] = result
-            
+
         return json, message, (end_page_no if end_page_no > 0 else n_page)
-    
-    def get_mobile(self, config: TableConfigDetail, data: dict, n_page: int, json: dict, message: ProcessMessages) -> Tuple[dict, ProcessMessages, int]:
+
+    def get_mobile(self, config: TableConfigDetail, data: dict, n_page: int, json: dict,
+                   message: ProcessMessages) -> Tuple[dict, ProcessMessages, int]:
         """
             Get the info from the table of the specific object
         Args:
@@ -120,17 +121,18 @@ class Davere01(DAvere):
             Tuple[dict, ProcessMessages,  int]: response JSON for the specific object, processing messages 
                                                     and the page number where the table ends
         """
-        
+
         tables, message, end_page_no = self.find_table_in_document_between_lines(data['ocr_form_response'], \
             n_page, self.no_of_pages, config.upper, config.lower, config.header, message)
-        
+
         message, result = self.extract_table_info_to_json('mobile', tables, lambda x: Mobile(), message)
         if message.has_errors() or result is not None:
             json['mobile'] = result
-            
+
         return json, message, (end_page_no if end_page_no > 0 else n_page)
-    
-    def get_finances(self, config: TableConfigDetail, data: dict, n_page: int, json: dict, message: ProcessMessages) -> Tuple[dict, ProcessMessages, int]:
+
+    def get_finances(self, config: TableConfigDetail, data: dict, n_page: int, json: dict,
+                     message: ProcessMessages) -> Tuple[dict, ProcessMessages, int]:
         """
             Get the info from the table of the specific object
         Args:
@@ -143,17 +145,18 @@ class Davere01(DAvere):
             Tuple[dict, ProcessMessages,  int]: response JSON for the specific object, processing messages 
                                                     and the page number where the table ends
         """
-        
+
         tables, message, end_page_no = self.find_table_in_document_between_lines(data['ocr_form_response'], \
             n_page, self.no_of_pages, config.upper, config.lower, config.header, message)
-        
+
         message, result = self.extract_table_info_to_json('finances', tables, lambda x: Finance(), message)
         if message.has_errors() or result is not None:
             json['finance'] = result
-            
+
         return json, message, (end_page_no if end_page_no > 0 else n_page)
-    
-    def get_investments(self, config: TableConfigDetail, data: dict, n_page: int, json: dict, message: ProcessMessages) -> Tuple[dict, ProcessMessages, int]:
+
+    def get_investments(self, config: TableConfigDetail, data: dict, n_page: int, json: dict,
+                        message: ProcessMessages) -> Tuple[dict, ProcessMessages, int]:
         """
             Get the info from the table of the specific object
         Args:
@@ -166,17 +169,18 @@ class Davere01(DAvere):
             Tuple[dict, ProcessMessages,  int]: response JSON for the specific object, processing messages 
                                                     and the page number where the table ends
         """
-        
+
         tables, message, end_page_no = self.find_table_in_document_between_lines(data['ocr_form_response'], \
             n_page, self.no_of_pages, config.upper, config.lower, config.header, message)
-        
+
         message, result = self.extract_table_info_to_json('investments', tables, lambda x: Investment(), message)
         if message.has_errors() or result is not None:
             json['investment'] = result
-            
+
         return json, message, (end_page_no if end_page_no > 0 else n_page)
-    
-    def get_extra_finance_info(self, data: dict, n_page: int, json: dict, message: ProcessMessages) -> Tuple[dict, ProcessMessages, int]:
+
+    def get_extra_finance_info(self, data: dict, n_page: int, json: dict,
+                               message: ProcessMessages) -> Tuple[dict, ProcessMessages, int]:
         """
             Get text from a specific section
         Args:
@@ -189,19 +193,20 @@ class Davere01(DAvere):
             Tuple[dict, ProcessMessages, int]: esponse JSON for the specific object, processing messages 
                                                     and the page number where the table ends
         """
-        
+
         lines, end_page_no = self.find_lines_in_document_between_lines(data['ocr_form_response'], \
                             n_page, self.no_of_pages, \
                             '3. Alte active producatoare de venituri nete,', None, False, \
                             'NOTA:', None, False)
-        
+
         result = self.extract_lines_info_to_json(lines)
         if result is not None and len(result) > 0:
             json['finance_extra_info'] = result
-            
+
         return json, message, (end_page_no if end_page_no > 0 else n_page)
-    
-    def get_debt(self, config: TableConfigDetail, data: dict, n_page: int, json: dict, message: ProcessMessages) -> Tuple[dict, ProcessMessages, int]:
+
+    def get_debt(self, config: TableConfigDetail, data: dict, n_page: int, json: dict,
+                 message: ProcessMessages) -> Tuple[dict, ProcessMessages, int]:
         """
             Get the info from the table of the specific object
         Args:
@@ -214,17 +219,18 @@ class Davere01(DAvere):
             Tuple[dict, ProcessMessages,  int]: response JSON for the specific object, processing messages 
                                                     and the page number where the table ends
         """
-        
+
         tables, message, end_page_no = self.find_table_in_document_between_lines(data['ocr_form_response'],  \
             n_page, self.no_of_pages, config.upper, config.lower, config.header, message)
-        
+
         message, result = self.extract_table_info_to_json('debt', tables, lambda x: Debt(), message)
         if message.has_errors() or result is not None:
             json['debt'] = result
-            
+
         return json, message, (end_page_no if end_page_no > 0 else n_page)
-    
-    def get_gift(self, config: TableConfigDetail, data: dict, n_page: int, json: dict, message: ProcessMessages) -> Tuple[dict, ProcessMessages, int]:
+
+    def get_gift(self, config: TableConfigDetail, data: dict, n_page: int, json: dict,
+                 message: ProcessMessages) -> Tuple[dict, ProcessMessages, int]:
         """
             Get the info from the table of the specific object
         Args:
@@ -237,18 +243,19 @@ class Davere01(DAvere):
             Tuple[dict, ProcessMessages,  int]: response JSON for the specific object, processing messages 
                                                     and the page number where the table ends
         """
-        
+
         tables, message, end_page_no = self.find_table_in_document_between_lines(data['ocr_form_response'],  \
             n_page, self.no_of_pages, config.upper, config.lower, config.header, message)
-        
+
         message, result = self.extract_table_info_one_level_to_json(tables, \
             config.first_level, lambda x: Gift(), message)
         if message.has_errors() or result is not None:
             json['gift'] = result
-            
+
         return json, message, (end_page_no if end_page_no > 0 else n_page)
-    
-    def get_income(self, config: TableConfigDetail, data: dict, n_page: int, json: dict, message: ProcessMessages) -> Tuple[dict, ProcessMessages, int]:
+
+    def get_income(self, config: TableConfigDetail, data: dict, n_page: int, json: dict,
+                   message: ProcessMessages) -> Tuple[dict, ProcessMessages, int]:
         """
             Get the info from the table of the specific object
         Args:
@@ -261,15 +268,14 @@ class Davere01(DAvere):
             Tuple[dict, ProcessMessages,  int]: response JSON for the specific object, processing messages 
                                                     and the page number where the table ends
         """
-        
+
         tables, message, end_page_no = self.find_table_in_document_between_lines(data['ocr_form_response'],  \
             n_page, self.no_of_pages, config.upper, config.lower, config.header, message)
-        
+
         message, result = self.extract_table_info_two_level_to_json(tables, \
             config.second_level, config.first_level, lambda x: Income(), message)
-        
+
         if message.has_errors() or result is not None:
             json['income'] = result
-            
+
         return json, message, (end_page_no if end_page_no > 0 else n_page)
-
